requirements.txt
google-api-python-client==2.118.0
pymongo==4.6.2
transformers==4.38.2
torch==2.2.1
google-generativeai==0.3.2
python-dotenv==1.0.1
pandas==2.2.1
numpy==1.26.4
requests==2.31.0
tqdm==4.66.2 